move to model_validator
